# artifacts/api-server/src/routes/reports.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { reportsTable, findingsTable, scansTable, workspacesTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import { requireAuth } from "../lib/session";
import { logAudit } from "../lib/audit";

const router = Router();

router.get("/workspaces/:workspaceId/reports", requireAuth, async (req, res) => {
  const reports = await db.select().from(reportsTable).where(eq(reportsTable.workspaceId, req.params.workspaceId)).orderBy(reportsTable.createdAt);
  res.json(reports);
});

router.post("/workspaces/:workspaceId/reports", requireAuth, async (req, res) => {
  const { title, type = "markdown" } = req.body;
  const wid = req.params.workspaceId;
  const [ws] = await db.select().from(workspacesTable).where(eq(workspacesTable.id, wid)).limit(1);
  if (!ws) { res.status(404).json({ error: "Workspace not found" }); return; }
  const [report] = await db.insert(reportsTable).values({ workspaceId: wid, title: title ?? `${ws.name} — Security Report`, type, status: "generating" }).returning();
  const findings = await db.select().from(findingsTable).where(eq(findingsTable.workspaceId, wid));
  const scans = await db.select().from(scansTable).where(eq(scansTable.workspaceId, wid));
  const sev = { critical: 0, high: 0, medium: 0, low: 0, info: 0 };
  for (const f of findings) { const s = f.severity as keyof typeof sev; if (s in sev) sev[s]++; }
  const content = `# ${report.title}\n\n**Workspace:** ${ws.name}\n**Target:** ${ws.target ?? "N/A"}\n**Generated:** ${new Date().toUTCString()}\n**Type:** ${ws.type}\n\n---\n\n## Executive Summary\n\nThis report covers the security assessment conducted against **${ws.target ?? ws.name}**.\nA total of **${findings.length} findings** were identified across **${scans.length} scans**.\n\n| Severity | Count |\n|----------|-------|\n| Critical | ${sev.critical} |\n| High     | ${sev.high} |\n| Medium   | ${sev.medium} |\n| Low      | ${sev.low} |\n| Info     | ${sev.info} |\n\n---\n\n## Findings\n\n${findings.map((f,i) => `### ${i+1}. ${f.title}\n\n**Severity:** ${f.severity.toUpperCase()}\n**Status:** ${f.status}\n**Component:** ${f.affectedComponent ?? "N/A"}\n${f.cve ? `**CVE:** ${f.cve}` : ""}\n${f.cvss ? `**CVSS:** ${f.cvss}` : ""}\n\n${f.description ?? "No description provided."}\n\n${f.recommendation ? `**Recommendation:** ${f.recommendation}` : ""}\n\n---`).join("\n\n")}\n\n## Scan Summary\n\n${scans.map(s => `- **${s.name}** (${s.tool}) — Status: ${s.status} — Target: ${s.target}`).join("\n")}\n\n---\n*Generated by GodsApp Security Suite*\n`;
  await db.update(reportsTable).set({ status: "completed", content, generatedAt: new Date() }).where(eq(reportsTable.id, report.id));
  const [final] = await db.select().from(reportsTable).where(eq(reportsTable.id, report.id)).limit(1);
  await logAudit("report.generated", { resource: "report", resourceId: report.id, req });
  res.status(201).json(final);
});

router.get("/workspaces/:workspaceId/reports/:reportId", requireAuth, async (req, res) => {
  const [report] = await db.select().from(reportsTable).where(and(eq(reportsTable.id, req.params.reportId), eq(reportsTable.workspaceId, req.params.workspaceId))).limit(1);
  if (!report) { res.status(404).json({ error: "Report not found" }); return; }
  res.json(report);
});

router.delete("/workspaces/:workspaceId/reports/:reportId", requireAuth, async (req, res) => {
  const [report] = await db.delete(reportsTable).where(and(eq(reportsTable.id, req.params.reportId), eq(reportsTable.workspaceId, req.params.workspaceId))).returning();
  if (!report) { res.status(404).json({ error: "Report not found" }); return; }
  await logAudit("report.deleted", { resource: "report", resourceId: req.params.reportId, req });
  res.json({ success: true });
});

export default router;