# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: GodsApp — Professional Security Auditing & Research Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/api-keys:
    get:
      operationId: getApiKeys
      tags: [settings]
      summary: Get all configured API keys (masked)
      responses:
        "200":
          description: API keys list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"

    put:
      operationId: upsertApiKey
      tags: [settings]
      summary: Create or update an API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "200":
          description: API key saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /settings/api-keys/{service}/test:
    post:
      operationId: testApiKey
      tags: [settings]
      summary: Test an API key connection
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyTestResult"
  /settings/general:
    get:
      operationId: getGeneralSettings
      tags: [settings]
      summary: Get general settings
      responses:
        "200":
          description: General settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettings"

    patch:
      operationId: updateGeneralSettings
      tags: [settings]
      summary: Update general settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneralSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettings"
...
  /workspaces/{workspaceId}/scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scans in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"

    post:
      operationId: createScan
      tags: [scans]
      summary: Start a new scan
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /workspaces/{workspaceId}/scans/{scanId}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan details and results
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetail"

    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Delete a scan
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/scans/{scanId}/stop:
    post:
      operationId: stopScan
      tags: [scans]
      summary: Stop a running scan
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /workspaces/{workspaceId}/scans/{scanId}/replay:
    post:
      operationId: replayScan
      tags: [scans]
      summary: Re-run a scan with the same or modified parameters
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: scanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayScanInput"
      responses:
        "201":
          description: New scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"

  # ── Findings ──────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports for a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a new report
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /workspaces/{workspaceId}/reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report details
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Scheduler ─────────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/schedules:
    get:
      operationId: listSchedules
      tags: [workspaces]
      summary: List scheduled scans for a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Schedules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"

    post:
      operationId: createSchedule
      tags: [workspaces]
      summary: Create a scheduled scan
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /workspaces/{workspaceId}/schedules/{scheduleId}:
    delete:
      operationId: deleteSchedule
      tags: [workspaces]
      summary: Delete a schedule
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

# ── Schemas ───────────────────────────────────────────────────────────────
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    ApiKey:
      type: object
      required: [service, configured, lastTested]
      properties:
        service:
          type: string
        configured:
          type: boolean
        maskedKey:
          type: ["string", "null"]
        lastTested:
          type: ["string", "null"]
        testStatus:
          type: ["string", "null"]
    ApiKeyInput:
      type: object
      required: [service, key]
      properties:
        service:
          type: string
        key:
          type: string
    ApiKeyTestResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        details:
          type: ["string", "null"]

    # ── Settings ────────────────────────────────────────────────────────────
    GeneralSettings:
      type: object
      required: [operatorName, defaultWorkspaceId, autoLockMinutes, defaultReportFormat, confirmDestructive]
      properties:
        operatorName:
          type: string
        organizationName:
          type: ["string", "null"]
        defaultWorkspaceId:
          type: ["string", "null"]
        autoLockMinutes:
          type: integer
        defaultReportFormat:
          type: string
        confirmDestructive:
          type: boolean
        autoSaveScanOutput:
          type: boolean
    GeneralSettingsUpdate:
      type: object
      properties:
        operatorName:
          type: string
        organizationName:
          type: ["string", "null"]
        defaultWorkspaceId:
          type: ["string", "null"]
        autoLockMinutes:
          type: integer
        defaultReportFormat:
          type: string
        confirmDestructive:
          type: boolean
        autoSaveScanOutput:
          type: boolean
...
    Scan:
      type: object
      required: [id, workspaceId, tool, target, status, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        tool:
          type: string
        target:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed, stopped]
        profile:
          type: ["string", "null"]
        options:
          type: object
          additionalProperties: true
        progress:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        resultSummary:
          type: ["string", "null"]
        findingCount:
          type: integer
    ScanInput:
      type: object
      required: [tool, target]
      properties:
        tool:
          type: string
        target:
          type: string
        profile:
          type: string
        options:
          type: object
          additionalProperties: true
    ScanDetail:
      type: object
      required: [scan, output, hosts]
      properties:
        scan:
          $ref: "#/components/schemas/Scan"
        output:
          type: array
          items:
            type: string
        hosts:
          type: array
          items:
            $ref: "#/components/schemas/ScanHost"
        ports:
          type: array
          items:
            $ref: "#/components/schemas/ScanPort"
        rawOutput:
          type: ["string", "null"]
    ScanHost:
      type: object
      required: [ip, status]
      properties:
        ip:
          type: string
        hostname:
          type: ["string", "null"]
        status:
          type: string
        os:
          type: ["string", "null"]
        macVendor:
          type: ["string", "null"]
        ports:
          type: array
          items:
            $ref: "#/components/schemas/ScanPort"
    ScanPort:
      type: object
      required: [port, protocol, state]
      properties:
        port:
          type: integer
        protocol:
          type: string
        state:
          type: string
        service:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        banner:
          type: ["string", "null"]
    ReplayScanInput:
      type: object
      properties:
        target:
          type: ["string", "null"]
        profile:
          type: ["string", "null"]
        options:
          type: object
          additionalProperties: true

    # ── Findings ────────────────────────────────────────────────────────────
...
    Report:
      type: object
      required: [id, workspaceId, type, format, status, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        type:
          type: string
          enum: [engagement, findings, evidence, scan, threat_intel, compliance, comparison, daily]
        format:
          type: string
          enum: [pdf, markdown, docx, xlsx, html, json, sarif]
        status:
          type: string
          enum: [generating, ready, failed]
        title:
          type: ["string", "null"]
        findingCount:
          type: integer
        size:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        options:
          type: object
          additionalProperties: true
    ReportInput:
      type: object
      required: [type, format]
      properties:
        type:
          type: string
          enum: [engagement, findings, evidence, scan, threat_intel, compliance, comparison, daily]
        format:
          type: string
          enum: [pdf, markdown, docx, xlsx, html, json, sarif]
        title:
          type: string
        options:
          type: object
          additionalProperties: true

    # ── Scheduler ───────────────────────────────────────────────────────────
    Schedule:
      type: object
      required: [id, workspaceId, tool, target, cronExpression, enabled, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        tool:
          type: string
        target:
          type: string
        cronExpression:
          type: string
        enabled:
          type: boolean
        lastRun:
          type: ["string", "null"]
        nextRun:
          type: ["string", "null"]
        lastStatus:
          type: ["string", "null"]
        options:
          type: object
          additionalProperties: true
        createdAt:
          type: string
    ScheduleInput:
      type: object
      required: [tool, target, cronExpression]
      properties:
        tool:
          type: string
        target:
          type: string
        cronExpression:
          type: string
        options:
          type: object
          additionalProperties: true